--- a/yarn-project/archiver/package.json
+++ b/yarn-project/archiver/package.json
@@ -61,8 +61,6 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
-    "@swc/core": "^1.4.11",
-    "@swc/jest": "^0.2.36",
     "@types/debug": "^4.1.7",
     "@types/jest": "^29.5.0",
     "@types/lodash.groupby": "^4.6.9",