--- a/yarn-project/aztec/package.json
+++ b/yarn-project/aztec/package.json
@@ -59,6 +59,8 @@
   "types": "./dest/index.d.ts",
   "devDependencies": {
     "@jest/globals": "^29.5.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "@types/koa": "^2.13.6",
     "jest": "^29.5.0",