--- a/yarn-project/circuits.js/package.json
+++ b/yarn-project/circuits.js/package.json
@@ -45,6 +45,8 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "@types/lodash.chunk": "^4.2.7",
     "@types/node": "^18.7.23",