--- a/yarn-project/circuits.js/package.local.json
+++ b/yarn-project/circuits.js/package.local.json
@@ -1,0 +1,5 @@
+{
+  "scripts": {
+    "build": "yarn clean && yarn generate && tsc -b"
+  }
+}