--- a/yarn-project/cli/package.json
+++ b/yarn-project/cli/package.json
@@ -64,8 +64,6 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
-    "@swc/core": "^1.4.11",
-    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "@types/lodash.startcase": "^4.4.7",
     "@types/node": "^18.7.23",