--- a/yarn-project/end-to-end/Dockerfile
+++ b/yarn-project/end-to-end/Dockerfile
@@ -40,7 +40,7 @@ RUN rm -rf /usr/src/noir-projects /usr/src/l1-contracts
 
 # Create minimal image.
 FROM node:18.19.1-slim
-RUN apt-get update && apt-get install jq gnupg wget -y && \
+RUN apt-get update && apt-get install jq gnupg wget netcat-openbsd -y && \
     wget --quiet --output-document=- https://dl-ssl.google.com/linux/linux_signing_key.pub | gpg --dearmor > /etc/apt/trusted.gpg.d/google-archive.gpg && \
     sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list' && \
     apt-get update && \