--- a/yarn-project/end-to-end/package.json
+++ b/yarn-project/end-to-end/package.json
@@ -64,7 +64,6 @@
     "resolve-typescript-plugin": "^2.0.1",
     "stream-browserify": "^3.0.0",
     "string-argv": "^0.3.2",
-    "ts-jest": "^29.1.0",
     "ts-loader": "^9.4.4",
     "ts-node": "^10.9.1",
     "tslib": "^2.4.0",