--- a/yarn-project/end-to-end/package.json
+++ b/yarn-project/end-to-end/package.json
@@ -44,6 +44,8 @@
     "@aztec/world-state": "workspace:^",
     "@jest/globals": "^29.5.0",
     "@noble/curves": "^1.0.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/fs-extra": "^11.0.2",
     "@types/jest": "^29.5.0",
     "@types/koa": "^2.13.9",