--- a/yarn-project/end-to-end/package.json
+++ b/yarn-project/end-to-end/package.json
@@ -81,6 +81,8 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "concurrently": "^7.6.0",
     "jest": "^29.5.0",