--- a/yarn-project/end-to-end/scripts/docker-compose-p2p.yml
+++ b/yarn-project/end-to-end/scripts/docker-compose-p2p.yml
@@ -37,6 +37,10 @@ services:
       P2P_NAT_ENABLED: 'false'
       P2P_ENABLED: 'true'
       BOOTSTRAP_NODES: '/ip4/p2p-bootstrap/tcp/40400/p2p/12D3KooWGBpbC6qQFkaCYphjNeY6sV99o4SnEWyTeBigoVriDn4D'
-    command: ${TEST:-./src/e2e_p2p_network.test.ts}
+    entrypoint: >
+      sh -c '
+        while ! nc -z sandbox 8080; do sleep 1; done;
+        yarn test ${TEST:-./src/e2e_deploy_contract.test.ts}
+      '
     volumes:
       - ../log:/usr/src/yarn-project/end-to-end/log:rw