--- a/yarn-project/entrypoints/package.json
+++ b/yarn-project/entrypoints/package.json
@@ -53,6 +53,8 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "jest": "^29.5.0",
     "ts-node": "^10.9.1",