--- a/yarn-project/ethereum/package.json
+++ b/yarn-project/ethereum/package.json
@@ -31,6 +31,8 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "@types/node": "^18.14.6",
     "jest": "^29.5.0",