--- a/yarn-project/foundation/package.json
+++ b/yarn-project/foundation/package.json
@@ -92,6 +92,8 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
+    "@swc/core": "^1.4.11",
+    "@swc/jest": "^0.2.36",
     "@types/bn.js": "^5.1.3",
     "@types/debug": "^4.1.7",
     "@types/detect-node": "^2.0.0",