--- a/yarn-project/key-store/package.json
+++ b/yarn-project/key-store/package.json
@@ -45,8 +45,6 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
-    "@swc/core": "^1.4.11",
-    "@swc/jest": "^0.2.36",
     "@types/jest": "^29.5.0",
     "@types/node": "^18.7.23",
     "jest": "^29.5.0",