--- a/yarn-project/simulator/src/public/index.ts
+++ b/yarn-project/simulator/src/public/index.ts
@@ -1,7 +1,7 @@
 export * from './db.js';
 export {
-  PublicExecution,
-  PublicExecutionResult,
+  type PublicExecution,
+  type PublicExecutionResult,
   isPublicExecutionResult,
   collectPublicDataReads,
   collectPublicDataUpdateRequests,