`ts-jest` is slow. It performs full type checking before we even hit a line of code. This can lead to upwards of 7+ seconds of wait time in some cases (like e2e tests).

It's not really necessary to do such type-checking when in a development cycle, as:
1. Your IDE is likely running tsc.
2. You're likely running an `yarn tsc -b --watch` in the root of the project, to ensure you're getting incremental builds applied to `dest` folders for cross module imports.
3. CI will do a full build as a backstop for devs that didn't do incrementals and forgot to sanity check the project with a `yarn build`.

This PR swaps `ts-jest` for `@swc/jest`, a rust transpiler which can leverage the explicit `type` imports in the previous PR to do very fast ts->js transpilation.

I think swapping out type-safety in this point of the dev cycle is worth the improved feedback cycle. Especially given points above.

It also knocks about a second off startup time by not calling `yarn bin jest` in the test script, as yarn startup itself is annoyingly slow.

Running foundation tests:
* before: `7.8s`
* after: `3.7s`

Running circuits.js tests:
* before: `13.3s`
* after: `5.8s`

I expect the performance improves even more as you higher up the stack and ts-jest had more to transpile.